docs: Fix "an user" typo in README.md
In english, the word "an" goes before words that begin with a vowel sound. Exceptions are expressions like "a user", "a unicorn" or "a union". Although they begin with the vowel "u", it has a consonant sound. It is also possible to have a consonant as the first letter and still use "an" instead of "a". For example: "an hour".